Unify Pelican import with the shared process_table path
